Don't create ExpressionEXPR just to unwrap it again.
Several methods in ExpressionTreeCallRewriter that return ExpressionEXPR
are called by both VisitCALL and GetExpression. The former return them
directly while the latter unwraps the contained Expression again.

Flip this, so that these methods return Expressions and they are only
wrapped in ExpressionEXPR if necessary.

Also rename ExpressionEXPR to ExpressionExpr and remove an override that
just calls the base.