Root NetEventSource used by System.Net.*
All of the networking tests seem to have some dependencies on the
NetEventSource that are inconsistent by platform.  I suspect that only
some platforms are rooting the members the test expects and the tests
are resilient to not getting events but not resilient to missing
metadata.

For now I'm rooting all of these where I've seen failures.  The tests
should be cleaned up / conditioned appropriately.

For System.Net.Http the library logs using anonymous types then the
tests reflect on the public properties over those types.  The getters on
these properties were trimmed so I've rooted all anonymous types in the
assembly.