Fix incremental build of S.R.CS.Unsafe
This project had a hack to run TargetsTriggeredByCompilation
unconditionally.

That was OK when all those did was run APICompat, but now that
we're rewriting the assemblies this is broken: it reruns trimming
on rebuild and tries to resign the already signed assembly as well.

This hack is no longer needed as of https://github.com/dotnet/buildtools/commit/2c3aea5119702df1ba8af98e7c9a0fb9669d0106 so removing.