Add enhanced overloads for RSA and DSA.Create
During the design of the ECC import/export feature we concluded that
better overloads for the asymmetric algorithms' Create methods should
exist which allow the user to specify the generation space (an ECCurve for ECC,
a keysize for RSA and DSA) or to specify key parameters for importing.

These methods are currently pure convenience, but in the future they could
end up using the arguments to change a preferred back-end provider;
for example, on netfx RSA.Create() could prefer RSACng when the Exponent is
larger than 4 bytes, since RSACryptoServiceProvider cannot process those keys.