Fix MaxResponseHeadersLengthSemantics (#17796)
* Fix MaxResponseHeadersLengthSemantics

The new HttpClientHandler.MaxResponseHeadersLength property is defined differently from the similarly
named WebRequestHandler and HttpWebRequest properties. The units are supposed to be in Kilobytes
(1024 bytes). This problem was discovered while doing .NET Core vs. .NET Framework compat testing.

Fix the semantics to match the existing WebRequestHandler and HttpWebRequest properties. Adjust tests.

Contributes to #17691

* Address PR feedback