Add build option to only build stuff that is part of NETCoreApp
pass /p:BuildNETCoreAppOnly=true will filter the product build
to only include things in corefx that are part of netcoreapp (i.e
shared framework).

Note this filtering doesn't work correclty for test projects because
some of them require things outside the shared framework to build and
run. We could extend it to test projects at some point if we want
to clean up those dependencies, but for now it is only intended
for build.cmd/sh