Add back package for System.Reflection.TypeExtensions (#17831)
Fixes #17819

System.Reflection.TypeExtensions is not part of NETStandard.Library.
It's included in NetCore.App, but the package is required in order to
ship the netstandard ref and desktop implementation.

We must ship the higher version of API (4.1.x) on net461 now that
net461 supports netstandard2.0.  Previously this API was only supported
starting with net462.