Remove function CryptoNative_GetStreamSizes
This function hasn't been called since (before) .NET Core 1.0 RTM. It was left in to allow for the
RTM binary to replace a 1.0-rc2 binary without causing exceptions.

Since there's no longer an upgrade concern for 1.0-rc2 customers to current versions of the binary,
let's go ahead and remove this dead code.