Store _pendingReadRequest into a local for multiple accesses
It should not be possible by construction for _pendingReadRequest to become null in SignalComplete after we check it for nullness, as all access to it is protected by a lock.  But for good measure, store it into a local and use that local for subsequent accesses.  If nothing else, we can avoid an extra field access.