Enable the creation of certificates and cert requests
The new CertificateRequest class allows a caller to specify a subject name,
public key, and extensions for an extension request into a PKCS#10 blob to send
to a certificate authority.

The new CertificateRequest class allows a caller to set all of the fields in an
X.509 v3 certificate, with the exception of the v2 IssuerUniqueId and SubjectUniqueId
fields, as they are considered deprecated.  Accelerators exist for self-signing,
as well as chain-signing given an existing X509Certificate2 instance with a
private key already associated.

This change also adds
CopyWithPrivateKey(this X509Certificate2 cert, {RSA|DSA|ECDSA} key)
to enable scenarios where a public cert and private key are held independently
but a unification is desired.  This makes UNIX-style .pem/.key loading easier,
but still requires users understand how to load the .key file.