Disable illink on alpine
The Alpine CLI was hacked together with the wrong bits for the shared
framework.  It used 1.0 assemblies and put them in the 1.1 folder.

As a result this 1.1 targeting app fails to bind when it hits an
assembly (System.Xml.ReaderWriter) that was different.

Workaround it temporarily by turning off illink for alpine.