Fix portable System.Security.Cryptography.Native EC2M issue
We are building the portable dotnet core on the oldest Linux distro we
support, which is currently RHEL7.2. But the OpenSSL on that distro
doesn't support the EC2M functions and so it effectively disables them
for the portable dotnet core on all Linux distros.
This change fixes it by adding copies of declarations of the five
missing functions to opensslshim.h if they are not found at the compile
time. That enables the shim to be able to use those functions on all
distros that support them.