Don't build packages in multiple legs when OfficialBuildId is set
Package publishing is tripping up on having multiple legs produce the
same package ID.

Here's a summary of all the packages we produce:
1. Framework identity packages are produced in all framework specific
legs and the -allConfigurations leg.
2. Framework runtime packages are produced in runtime+framework-specific
legs and the -allConfigurations leg.
3. Standalone library packages are produced in the -allConfigurations
leg.

To ensure a single instance in official builds we need to change this as
follows:

1. Framework identity packages should only be produced in the
-allConfigurations leg.
2. Framework runtime packages should only be produced in the runtime+
framework-specific leg, not the -allConfigurations leg.