Make `netfx` mean `net461` instead of `net463`
We're moving the desktop .NETStandard2.0 support library build into
corefx so that we can build partial facades and implementation
assemblies and keep it compatible with previous netstandard versions.

Since we need to make net461 support netstandard2.0 this will become the
primary target for desktop builds in this repo.  As a result we need to
remap what netfx means.