Share build traversal logic across framework pkgs
Previously the build traversal logic was duplicated between the UAP and
NETCoreApp packages.

This refactors it into shared targets/props.

The mechanism for opt-in to this traversal behavior is to include a
props file in the project that matches the convention *.rids.props.
The file must define PackageRID and OfficialBuildRID items.