Create an init-tools.msbuild project file.
This file is used in init-tools.cmd/sh, and replaces the "generated"
project file used previously. Because we are using MSBuild projects for
nuget restore now, we are able to "parameterize" the version of
Microsoft.DotNet.BuildTools based on an MSBuild property which is passed
in to the "dotnet restore" command in init-tools.