Add netfx configurations for libs in NETStandard
For any libary that is part of netstandard and needs to either add types
to desktop or satisfy a higher assembly version that we previously
shipped, ensure we have both a netfx ref and src build.

The ref build is required because previously we were using the src build
(partial facade) as both ref and lib.  This is problematic because we
need to build the full netstandard.dll facade before building src
projects.  Ensuring a ref build also helps reduce package size since we
don't end up shipping the partial facade implementation in ref.

Src build has been added where it wasn't already present to fill in the
missing types, members, or satisfy the previously shipped assembly
version in netstandard1.x.