Fix Begin/EndSendFile on macOS
Begin/EndSendFile are currently hanging on macOS when the sendfile operation can't complete synchronously. Unlike sendfile on Linux, the sendfile variant on macOS can return with an error of EAGAIN or EINTR but having sent some of the data.  And it appears it's quite common for this to happen with EAGAIN, as it fills the send buffer and then returns EAGAIN if any data remains after that.

Currently the implementation isn't counting the data that was sent when one of these "errors" occurs, resulting in hangs as we don't believe we've sent everything we should have.  This commit fixes the implementation to appropriately track data transferred even during EAGAIN or EINTR failures.