Fix newly added GetProcessesByName_NoSuchProcess_ReturnsEmpty test
On macOS, we may be able to discover that a process exists but not be able to get some detail about it, like its name, if the process is running at a higher privilege level.  As a result, the GetProcessesByName_NoSuchProcess_ReturnsEmpty frequently fails in CI, as it expects to find no processes with an empty name but does.