Fix zero-length async receives on macOS
macOS appears to special-case a zero-length non-blocking receive call to complete successfully immediately even if there's no data available.  This breaks a commonly used pattern that works on both Windows and Linux of using a zero-length ReceiveAsync to be asynchronously notified of data being available.  As a fix, we special-case a zero-length ReceiveAsync to check whether any data is available as a stand-in for a receive, simulating EAGAIN if no data is currently available.