Fix FileVersionInfo on Unix for non-regular files
FileVersionInfo.GetVersionInfo on Unix attempts to read from the file in order to parse out the metadata stored in it.  This is only intended to work with regular files.  If it instead is given something non-regular, like a named pipe, it could hang trying to read from it, and regardless it won't successfully parse any metadata.

The fix is to change GetVersionInfo to detect the file type and throw for anything other than files.  For the test project, because the test involved adding a P/Invoke that won't work in UWP, I've split the test project into a Windows build and a Unix build.