Remove dead check from EmitBinaryMethod.
A method-based binary operator will be constructed in
GetMethodBasedBinaryOperator. If the type is lifted then it
must either have the return type lifted (which means IsLiftedToNull
will be true) or have that return type be bool.
Hence in the path for IsLifted but !IsLiftedToNull the test
(mc.Type != typeof(bool) can never pass.

Remove it.