Fix attribute handling of hidden directories ending with slashes
We currently consider a file to be hidden if its name starts with a period.  That applies to directories as well.  But if the directory name in the FileSystemInfo ends with a '/', then the file name that comes back from GetFileName is empty, and it's actually the GetDirectoryName value we want to check for a period.