Workaround cycles in IsolatedStorage
IsolatedStorage doesn't cross-compile for unix and uses Windows-specific
libraries.

These libraries have a dummy implementation that throw
PlatformNotSupported on Unix.  This dummy implementation was compiled
against netstandard.dll.

This is a problem since IsolatedStorage is part of the closure of
netstandard.dll.  We correctly catch these cases as a cycle.

Workaround this by making netcoreapp-Unix-specific impls of these that
compile against System.*, putting them lower in the stack.
https://github.com/dotnet/corefx/issues/18578