S.L.Expressions: Consolidate pointer & byref test in ValidateType
The majority of calls to ValidateType happen along with tests that
the type validated isn't a pointer or byref (in the other direction,
all pointer checks and the majority of byref checks happen this way).

Make these checks part of ValidateType with an overload for the
minority case where they aren't necessary.