Fix several System.IO.Packaging issues with differing desktop behavior
On desktop, calling Package.Open(stream, FileMode.Create, ...) on a seekable stream sets the length to 0, since the ZipArchive that gets created from it will seek around in the stream.  The core implementation was not doing that.  Then tests were written against core; those tests should have used FileMode.Open but used FileMode.Create (presumably by accident), and thus on desktop the stream was being cleared whereas on core it was not.

This change fixes the core implementation to do the same stream validation/mutation that's done on desktop, and updates the tests to use FileMode.Open instead of FileMode.Create in the few problematic cases.