Fix System.Net.NameResolution tests on desktop
A bunch of the unit tests try to validate that sockets were correctly initialized.  They do so with fakes that intercept the calls.  But on desktop, the real methods were being used, bypassing that interception.  I've fixed it by making the unit tests on desktop compile like core so as to use the fakes.