Make desktop ns2.0 support package support net47
We need to support differing ref/impl per version of desktop until it
adds inbox support for ns2.0.

To do this I've made the netfx build also build for net462 and net47.

Then I've moved the files into framework-specific folders for each TFM.

For a given TFM it will prefer the files in its own folder then fallback
to the previous TFM if it doesn't have that file.