Fix clean of binplaced files/props
I noticed that incremental clean was broken for binplaced files. This
was due to use of batching outside a target.  I've fixed that by using
a transform instead.

I also noticed that /t:clean wasn't working at all.  This is due to
clean using a different target: coreclean.  I've added a target that
extends this as well.

Finally I noticed that we were never cleaning the generated props files.
This was because I forgot to add them to file writes and we didn't
consider them as additional directories.