Remove UAPAOT configuration of Vectors.WinRT
This library doesn't have a ref so its lib is used for both ref and lib.
That's causing a problem since the library also builds for two
configurations (UAP and UAP AOT) which both binplace refs to the same
folder.

Fix this by removing the UAPAOT configuration.  It wasn't used anyway.