Fix building with only VS2017 installed.
We weren't passing the right setting to CMAKE.

Also fixes issue with current working directory getting changed during the build. Run.exe currently must be run
with the current working directory set to the repo root. One part of this was the config.json file- which I've explicitly specified now.

The other (as of yet unfixable) problem is that we specify the location of the binclash logger as a relative path. Afaik there currently is no way to make this happen with run.cmd without making this logger an explicit  (as opposed to default) option. Could possibly extend run.exe's parsing to allow specifying environment variables or reference to other parameters...

I haven't nailed down who is changing the working directory- while it appears to be a 2017 specific problem I can't guarantee that it isn't some other variable.