Allow HttpWebRequest.AllowReadStreamBuffering to be set to true
While fixing various .NET Framework app-compat differences, PR #18163 changed the
AllowReadStreamBuffering property to match .NET Framework. This resulted in having
the property throw an exception when set to true. That behavior is what .NET Framework
currently does.

Historically, this property did not always exist in .NET Framework at all. Other portable
versions of .NET (Windows Phone, Silverlight) for the HTTP stack were implemented differently
and had this property which worked. When .NET Framework APIs were reconciled in 2015, this
virtual property was added to .NET Framework but the functionality on Desktop was limited.

This PR will put back this property's functionality for .NET Core which is also used for UWP apps.

Fixes #18668.