Validate that issuerCertificate is a sensible CA certificate
When a certificate is given as a CA cert it needs to
* Not have a KeyUsage extension, or assert KeyCertSign.
* Have BasicConstraints.CA set.

These two rules are a discovered state of Windows chain validation after
cross-reading RFC 3280, RFC 5280 and ITU-T X.509 2012/10.

In addition to the CA cert sanity checks, enforce that the notBefore and
notAfter values supplied are aligned with the issuerCertificate's NotBefore
and NotAfter values.