Make sure targetgroups>=net461 import netstandard
This is required for the vertical build.  NETStandard needs to be
synonymous with netstandard2.0 in order to get the correct fallback
behavior.  This isn't how its modeled, it's actually a different
distinct node in the compatibility graph.  While that's true, we need
to make sure we use it.

I plan to fix this better with https://github.com/dotnet/corefx/issues/18736