Trap tests that start child processes on AOT.
A number of tests use the corefx RemoteExecutorTestBase
helper to start child processes.

Since this helper has never heard of .NET Native,
it ends up through a series of accidents, reinvoking
xunit.console.netcore.exe recursively with a
command line that xunit.console.netcore.exe has
no clue of understanding.

We'll now properly detect this case and throw
an informative exception for future run-ins
with this.