Make the OSX mutable store tests skip on OSStatus -61.
Something about the hosting model in Helix is causing it to always fail with
errSecWrPerm (-61).  I can reproduce this problem by making ~/Library/Keychains readonly; and I can produce other problems using ssh and
sudo, but I can't quite reproduce a hosting model problem which produces
errSecWrPerm.

So "run" one of the tests, and if it fails with CryptographicException(-61) then
don't run the tests.  If it fails for any other reason, run the tests to get the
telemetry.