Simplify X509Chain building with OpenSSL
Experience since this code was written has shown the places were we worked
too hard to build the chain.

The `store` parameter to X509_STORE_CTX_init contains CRLs (which we did)
and the certificates which are to be treated as trust anchors.  Since we put
everything in this set we then had to go back and independently determine if
a trusted root capped off the chain.

The `chain` parameter contains certificates that are "untrusted". Windows has
a cert store named "Distrusted", which is what I long ago thought this meant,
what it means is "not trusted", or "of unknown provenance".

By correctly partitioning the certificates we can save extra work that OpenSSL
was already doing for us.

Any cert errors which were only determined once the whole chain was built
were reported on all the certs, instead of just the one that matched it. This
is due to the verify callback being called with success (ok=1) on every cert
at the end of the chain walk/verify, but the last error is still resident in the
last error field.  So only process error codes when called in failure mode (ok=0).