Simplify X509Chain building with OpenSSL  (#19017)
While looking into why some chain errors show on all certificates, vs just the ones where the problem originates, I ran across some information that would have simplified the work of building chains long ago.

By providing the candidates set (including the values from ChainPolicy.ExtraStore) and the trusted store (the union of the CurrentUser\Root and LocalMachine\Root stores) in separate collections OpenSSL will tell us when the root was trusted or not. So we should do that, and get rid of our post-facto decision that says the root was not trusted after all.

The error codes getting copied in multiple positions was an artifact of the callback being called at the end with ok=1. Most of the time the X509StoreCtxGetError value has been cleared to X509_V_OK, but some of them remain. By only reading error codes when ok=0 they're not only placed on the element where we first read them.

There's a fair body of tests which already verify LM\Root trust; and I manually verified behaviors with CU\Root.