Remove SafeHandle marshaling from Send/ReceiveAsync operations on Windows
In high-throughput situations with asynchronous operations on Sockets, the overhead involved in marshaling SafeHandles is showing up as a non-trivial cost, contributing anywhere from 3-5% overhead.  While the SafeHandles help to avoid certain misuses, a) SafeHandle marshaling (and in particular the implicit use of DangerousAddRef/Release) already isn't being done for synchronous operations (presumably to enable operation cancelation by closing the handle concurrent with the operation), and it's only being done for asynchronous operations during the actual P/Invoke (and thus the SafeHandle isn't "protected" while the async operation is in flight).  The benefit for the cost isn't worth it.  This commit removes the marshaling for send/receive operations.