Disable HttpWebRequest Abort related test
Similar to #18800.

There is a race condition between starting the I/O with
BeginGetRequestStream() and calling Abort().  The test tends to work
more on .NET Core since a MemoryStream is returned as the request
stream. But on .NET Framework, the request stream can be network based.

The test is left disabled for now in the code.

Contributes to #19083