Fix using a generic type parameter on a method with DispatchProxy.
For return types, the assert was triggered in debug mode.

For parameters, an ExecutionEngineException was triggered at run time on
x64 Windows when passing a value type. The lack of an box instruction
causes an AV in JIT_Stelem_Ref when it tries to treat the value type as an
Object* .