Make ECDsa.Create(ECCurve(.IsExplicit)) work on Windows. (#19208)
ECDsa.Create(ECCurve) had a check which rightly belonged in GenerateKey
behind an IsNamed check.  Moved that check to where it belonged, and added
an Explicit (PrimeShortWeierstrass) test to the big create matrix.

Adding the test resulted in debugging several tests which were assuming IsNamed.
The extra debugging data seemed generally applicable, so it remains.