Fix WinHttpHandler POST behavior with zero-length data
WinHttpHandler was incorrectly sending two zero-length chunks (a violation of HTTP protocol)
when POST'ing empty data using chunked encoding. The second zero-length chunk was ok since it
means end of data. The first zero-length chunk was because it was the empty string being encoded.

This was causing problems by sending extra data to WinHTTP and hence to the server. Any
subsequent requests to the server might get mis-understood by the server and the server would then
send back an error. That translated to getting a WinHTTP error (12152)
"The server returned an invalid or unrecognized response".

This is hard to repro except in the HttpListener tests because of the timing of when the extra
zero-length chunk of request data gets sent to the server. There is already an HttpClient unit test
for POSTing zero-length data using chunked encoding. But this but was not triggering this error
due to timing issues. Even the HttpListener test needed to be adjusted to force a consistent repro.

Added asserts to validate the logic of making sure that we don't try to chunk encode zero length
data.

Fixes #18869.