Fix marshaling-related ArgumentException in ExpandEnvironmentVariablesCore
This bug requires a certain pattern of inputs and calls to manifest, e.g.
- StringBuilderCache contains a StringBuilder with Capacity == N, where N > 100
- Call ExpandEnvironmentVariablesCore on 32-bit with a string (no quoted regions) of length N+1

The ExpandEnvironmentVariablesCore implementation will acquire a StringBuilder of at least length 100, tracking the requested size of 100 in "currentSize" rather than the actual capacity it gets back.  Then it will make an initial call to ExpandEnvironmentStringsW with the StringBuilder and the tracked currentSize.  The length will be insufficient, so the Win32 call will fail and return the required size.  But the runtime marshaling for the StringBuilder doesn't know it failed, so it dutifully marshals back the results of the call, and in doing so, ends up increasing the capacity of the StringBuilder and, importantly, increasing its Length, such that the Length is now larger than both the required size and currentSize.  It will then try to increase the capacity of the builder to the required size, but the length and capacity are already longer than that, and the call to set the Capacity will throw due to it being perceived as losing data.

The short-term fix is simple: just clear the builder before rather than after changing its capacity.  A better longer-term fix is to stop using StringBuilder entirely; it adds overhead for little benefit.