Skip DualMode localhost tests when it's not
Several DualMode tests assume that "localhost" maps to both IPv4 and IPv6 addresses, either because the test is verifying that a DualMode socket (such as one used by Socket.ConnectAsync) is able to connect to both IPv4 and IPv6 endpoints, or because the test is verifying that a DualMode socket listening at a given address can accept both IPv4 and IPv6 connections.  but "localhost" doesn't always map to both IPv4 and IPv6.  Change the tests to be conditional to only run when that assumption holds.