Added constructors to immediately open an X509Store (#19145)
* Added new constructors to X509Store and IsOpen property

As discussed in #6550 two new constructor overloads have been added that immediately opens the store using the provided flags. Additionally an IsOpen property has been added that returns whether the store is open or not.

* Removed #ifdef's

* Use ThrowsAny to allow derived exception types

On Windows we throw a WindowsCryptographicException, which derives from CryptographicException. Now we're using ThrowsAny here to allow the test to pass on Windows as well.

* Added more tests and remove unnecessary this