Add missing "return 0;" to main in several check_cxx_source_compiles
Thanks to the PR feedback, I've noticed that about 10 of the
check_cxx_source_compiles in configure.cmake are missing return
statement at the end of the main function. While it doesn't
break the detection now, a more strict future version of the
clang might have a problem with it.