Ignore lack of CURLMOPT_MAX_HOST_CONNECTIONS in CurlHandler
Versions before 7.30 don't have this option.  Previously we would throw a PlatformNotSupportedException.  As devs would generally use an option of this type to increase the limit (since historically such a limit was 2 or 10 in the .NET Framework) rather than decrease it, and since it's already effectively infinite by default in .NET Core, with this commit, now we just ignore it.  This helps to avoid PNSEs on the few platforms that still ship with too old a version of libcurl.