Fix FindByValidThumbprint_RootCert when an invalid root is found.
One of our CI machines has a certificate marked as rootTrust which nevertheless
comes back as not trusted.  This makes the test a bit more resilient against
failures of that nature (though it does make it somewhat hard to know if it
found a system with no candidate certs, or if it actually worked).