Fix LazyTests.Ctor_ExceptionRecovery test
On AoT, XUnit itself was invoking the Lazy object's
Value property behind our backs. This particular
test is sensitive to that.

(The reason why this was happening on AoT
and not CoreClr was because CoreClr finds
Lazy's ToString() override and invokes that.
We could make the same happen pm AoT with an rd.xml
but it's simply more reliable not to
give XUnit a chance to meddle in the first
place.)