Block opening the Disallowed X509Store on Linux
On Linux we don't use the Disallowed store for making (anti-)trust decisions, but
since we didn't go out of our way to block it a user can reference it either by
name or the StoreName enum and we'll happily create a store of that name, it
just won't do what the user expects.

So we should throw if they try to open the store, so we're not construed as
implicitly making promises that we don't follow through on.