Handle the PNSE when opening LocalMachine\My store on Linux
When an X509Data is being used as a source of hunting for a store certificate
with a (matching) private key it fails on Linux with a
PlatformNotSupportedException when trying to open the LocalMachine\My
store.

Since the code was already resilient to CryptographicException from the store
open, add PlatformNotSupportedException to things it expects as ignorable.