Rewrite SendFile_APM test
The test sporadically failed with a timeout, but little details are available about why.  Most likely it's due to a failure in one of the operations that causes the other to hang.  It could also be due to errors getting eaten.  I've rewritten the test to both make it simpler code and to address these issues; if it fails again, we'll have a better idea of why.