Fix 40 or so System.Net.Mail.Functional.Tests failures on ILC (#19602)
* Fix 40 or so System.Net.Mail.Functional.Tests failures on ILC

SmtpClient class is trying to fill in the "client host name"
checkbox for the HELO message. By default, it tries to
use the local host name but this is PNS on .Net Native.
This one hiccup makes SmtpClient unusable.

Since the client host name is apparently discretionary anyway,
fall back to LocalHost if the actual local host is PNS (or NS - why
not) as we do in the garbled-Unicode case.

* Fix 40 or so System.Net.Mail.Functional.Tests failures on ILC

SmtpClient class is trying to fill in the "client host name"
checkbox for the HELO message. By default, it tries to
use the local host name but this is PNS on .Net Native.
This one hiccup makes SmtpClient unusable.

Since the client host name is apparently discretionary anyway,
fall back to LocalHost if the actual local host is PNS (or NS - why
not) as we do in the garbled-Unicode case.

* Fix 40 or so System.Net.Mail.Functional.Tests failures on ILC

SmtpClient class is trying to fill in the "client host name"
checkbox for the HELO message. By default, it tries to
use the local host name but this is PNS on .Net Native.
This one hiccup makes SmtpClient unusable.

Since the client host name is apparently discretionary anyway,
fall back to LocalHost if the actual local host is PNS (or NS - why
not) as we do in the garbled-Unicode case.